requirements.txt
grpcio==1.66.1
grpcio-tools==1.66.1
protobuf==5.28.1
pymongo==4.8.0
python-dotenv==1.0.1